move crypto_utils.ts into encryption package
it contains the aes gcm code that i’d like to use for channel messages
